+++ src/main.py
@@ -1,0 +1,4 @@
+import frontend.gui as fr
+
+if __name__ == "__main__":
+    fr.run()
